requirements.txt
Django>=5.0.1,<5.1.0
djangorestframework>=3.14.0,<3.15.0
drf-spectacular>=0.27.0,<0.28.0
freezegun>=1.4.0,<1.5.0
psycopg2>=2.9.0,<3.0.0
python-dotenv>=1.0.0,<1.1.0
